refactor: made hook for creating the articles
